+++ CHANGELOG.md
@@ -4,6 +4,7 @@ dev:
   - add deneb information to "block info"
   - add epoch parameter to "validator yield"
   - add proposer index to "block info"
+  - "block info" honours "--quiet" flag
 
 1.31.0:
   - initial support for deneb